requirements.txt
black==20.8b1
celery==5.0.5
Django>=3.0,<4.0
djongo==1.3.4
django-celery-beat==2.2.0
djangorestframework==3.12
factory-boy==3.2.0
Faker==6.6.3
flake8==3.9.0
isort==5.8.0
pyflakes==2.3.1
python-decouple==3.4
psycopg2-binary>=2.8